refactor: consolidate webpack configs
