+++ package.json
@@ -4,18 +4,12 @@
   "version": "0.5.0-alpha",
   "description": "desktop application for the lightning network",
   "scripts": {
-    "build": "cross-env NODE_CONFIG_ENV=production concurrently --raw \"npm:build-main\" \"npm:build-preload\" \"npm:build-renderer\"",
-    "build-e2e": "concurrently \"npm run build-main\" \"npm:build-preload\" \"npm run build-renderer-e2e\"",
-    "build-dll": "webpack -r @babel/register --config webpack/webpack.config.renderer.dev.dll.js",
-    "build-main": "webpack -r @babel/register --config webpack/webpack.config.main.prod.js",
-    "build-preload": "webpack -r @babel/register --config webpack/webpack.config.preload.prod.js",
-    "build-preload-dev": "webpack -r @babel/register --config webpack/webpack.config.preload.dev.js",
-    "build-renderer": "webpack -r @babel/register --config webpack/webpack.config.renderer.prod.js",
-    "build-renderer-e2e": "webpack -r @babel/register --config webpack/webpack.config.renderer.e2e.js",
+    "build": "cross-env NODE_CONFIG_ENV=production webpack -r @babel/register --config webpack/webpack.config.prod.js",
+    "build-e2e": "cross-env NODE_CONFIG_ENV=production webpack -r @babel/register --config webpack/webpack.config.e2e.js",
+    "build-dll": "webpack -r @babel/register --config webpack/webpack.config.dll.js",
     "clean": "rimraf node_modules dll dist coverage .eslintcache",
     "coverage": "open coverage/index.html",
     "coveralls": "cat ./coverage/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
-    "predev": "npm run build-preload-dev",
     "dev": "cross-env HOT=1 npm run start-renderer-dev",
     "extract-messages": "extract-messages -l=af-ZA,ca-ES,de-DE,es-ES,ga-IE,hr-HR,ja-JP,no-NO,pt-PT,sr-SP,uk-UA,zh-TW,ar-SA,cs-CZ,el-GR,fi-FI,he-IL,hu-HU,ko-KR,pl-PL,ro-RO,sv-SE,vi-VN,bg-BG,da-DK,en,fr-FR,hi-IN,it-IT,nl-NL,pt-BR,ru-RU,tr-TR,zh-CN -o translations -d en --flat true renderer/**/messages.js",
     "fetch-lnd": "node ./scripts/fetch-lnd-for-packaging.js",
@@ -36,7 +30,7 @@
     "prestart": "cross-env NODE_CONFIG_ENV=production npm run build",
     "start": "cross-env NODE_ENV=production electron .",
     "start-main-dev": "electron -r @babel/register ./electron/main.js",
-    "start-renderer-dev": "webpack-dev-server -r @babel/register --config webpack/webpack.config.renderer.dev.js  --progress",
+    "start-renderer-dev": "webpack-dev-server -r @babel/register --config webpack/webpack.config.dev.js  --progress",
     "storybook": "cross-env NODE_ENV=storybook start-storybook -p 6006",
     "storybook:build": "cross-env NODE_ENV=storybook build-storybook",
     "storybook:deploy": "npm run storybook:build && gh-pages -t -d storybook-static -o origin -b gh-pages",
@@ -375,5 +369,5 @@
     "window-or-global": "1.0.1",
     "yup": "0.27.0"
   },
-  "main": "./dist/main.prod.js"
+  "main": "./dist/main.js"
 }
