+++ webpack/dev/main.config.js
@@ -1,47 +1,40 @@
 /**
- * Webpack config for development electron preload process
+ * Webpack config for production electron main process
  */
 
 import path from 'path'
 import merge from 'webpack-merge'
 import { EnvironmentPlugin } from 'webpack'
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer'
-import baseConfig, { rootDir } from './webpack.config.base'
+import baseConfig, { rootDir } from '../webpack.config.base'
+import devServer, { publicPath } from './common/devserver'
 
-export default merge.smart(baseConfig, {
-  devtool: 'source-map',
-
-  target: 'electron-renderer',
-
+const config = merge.smart(baseConfig, {
+  name: 'main',
+  target: 'electron-main',
   mode: 'development',
-
+  devtool: 'source-map',
   entry: {
-    preload: path.join(rootDir, 'electron', 'preload'),
+    main: path.join(rootDir, 'electron', 'main'),
   },
-
   output: {
-    path: path.join(rootDir, 'dist'),
-    filename: '[name].dev.js',
+    filename: '[name].js',
+    publicPath,
   },
-
+  devServer,
   plugins: [
     new EnvironmentPlugin({
       NODE_ENV: 'development',
     }),
-
     new BundleAnalyzerPlugin({
       analyzerMode: process.env.OPEN_ANALYZER === 'true' ? 'server' : 'disabled',
       openAnalyzer: process.env.OPEN_ANALYZER === 'true',
     }),
   ],
-
-  /**
-   * Disables webpack processing of __dirname and __filename.
-   * If you run the bundle in node.js it falls back to these values of node.js.
-   * https://github.com/webpack/webpack/issues/2010
-   */
   node: {
     __dirname: false,
     __filename: false,
   },
 })
+
+export default config
