+++ webpack/prod/main.config.js
@@ -6,44 +6,33 @@ import path from 'path'
 import merge from 'webpack-merge'
 import { EnvironmentPlugin } from 'webpack'
 import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer'
-import baseConfig, { rootDir } from './webpack.config.base'
+import baseConfig, { rootDir } from '../webpack.config.base'
 
-export default merge.smart(baseConfig, {
-  devtool: 'source-map',
-
+const config = merge.smart(baseConfig, {
+  name: 'main',
   target: 'electron-main',
-
   mode: 'production',
-
-  externals: ['config'],
-
+  devtool: 'source-map',
   entry: {
     main: path.join(rootDir, 'electron', 'main'),
   },
-
   output: {
+    filename: '[name].js',
     path: path.join(rootDir, 'dist'),
-    filename: '[name].prod.js',
   },
-
   plugins: [
     new EnvironmentPlugin({
-      NODE_ENV: 'production',
+      NODE_ENV: 'development',
     }),
-
     new BundleAnalyzerPlugin({
       analyzerMode: process.env.OPEN_ANALYZER === 'true' ? 'server' : 'disabled',
       openAnalyzer: process.env.OPEN_ANALYZER === 'true',
     }),
   ],
-
-  /**
-   * Disables webpack processing of __dirname and __filename.
-   * If you run the bundle in node.js it falls back to these values of node.js.
-   * https://github.com/webpack/webpack/issues/2010
-   */
   node: {
     __dirname: false,
     __filename: false,
   },
 })
+
+export default config
